+++ R/output_format.R
@@ -778,6 +778,11 @@ citeproc_required <- function(yaml_front_matter,
   ) && (
     !is.null(yaml_front_matter$bibliography) ||
       !is.null(yaml_front_matter$references) ||
-      length(grep("^references\\:\\s*$", input_lines)) > 0
+      # detect references: and bibliography: outside of yaml header
+      # as Pandoc is supporting
+      # TODO: remove when supporting multiple yaml block
+      # https://github.com/rstudio/rmarkdown/issues/1891
+      length(grep("^references:\\s*$", input_lines)) > 0 ||
+      length(grep("^bibliography:\\s*$", input_lines)) > 0
   )
 }