+++ NEWS.md
@@ -13,11 +13,14 @@ rmarkdown 2.6
 
 - Fix `pandoc_convert(citeproc = TRUE)` not supressing the `--natbib` or `--biblatex` options (thanks, @atusy, #1932).
 
+- `pandoc-citeproc` is now activated if a `bibliography` field is defined in another YAML block instead of the first YAML block (thanks, @bwiernik, #1364).
+
 - Specify that `htmltools::htmlPreserve()` should use the pandoc raw attribute 
   rather than preservation tokens when pandoc >= v2.0. Note that this option will
   have the intended effect only for versions of htmltools >= 0.5.0.9003.
 
 - `anchor_sections` in `html_documents()` now defaults to `FALSE`. It was introduced in previous version with a default to `TRUE`, but it is reverted now after hearing feedbacks from the community (thank you!). The `#` is still used as the character for the anchor but you can easily change that using CSS rules. Examples have been added to the help page `?html_document`.
+
 
 rmarkdown 2.5
 ================================================================================